moved Map component into designated folder
