+++ frontend/src/App.jsx
@@ -1,4 +1,4 @@
-import Map from "./components/Map";
+import Map from "./components/Map/Map";
 
 function App() {
   return (
